Crate psd

Data structures and methods for working with PSD files.

You are encouraged to read the PSD specification before contributing to this codebase. This will help you better understand the current approach and discover ways to improve it.

psd spec: https://www.adobe.com/devnet-apps/photoshop/fileformatashtml/

Structs

Psd
Represents the contents of a PSD file

PsdLayer
Information about a layer in a PSD file.

Enums

ColorMode
Adobe Docs

PsdChannelCompression
Indicates how a channe'sl data is compressed

PsdChannelKind
The different kinds of channels in a layer (red, green, blue, ...).

PsdDepth
Adobe Docs